Refactor package layout under src directory
