+++ src/ssl4polyp/__init__.py
@@ -1,0 +1,8 @@
+"""Top-level package for SSL4POLYP."""
+
+__all__ = [
+    "classification",
+    "models",
+    "polypdb",
+    "utils",
+]
