+++ src/ssl4polyp/classification/__init__.py
@@ -1,0 +1,1 @@
+"""Classification workflows for SSL4POLYP."""
