+++ src/ssl4polyp/classification/data/__init__.py
@@ -1,0 +1,1 @@
+"""Data loading utilities for classification experiments."""
