+++ src/ssl4polyp/classification/data/dataset.py
