+++ src/ssl4polyp/classification/data/manifest_dataset.py
