+++ src/ssl4polyp/classification/manifests.py
