+++ src/ssl4polyp/classification/metrics/__init__.py
@@ -1,0 +1,1 @@
+"""Metrics for evaluating classification models."""
