+++ src/ssl4polyp/models/__init__.py
@@ -1,0 +1,1 @@
+"""Model architectures used by SSL4POLYP."""
