+++ src/ssl4polyp/models/mae/__init__.py
@@ -1,0 +1,1 @@
+"""Masked autoencoder models and utilities."""
