+++ src/ssl4polyp/models/mae/util/__init__.py
@@ -1,0 +1,1 @@
+"""Utility helpers for masked autoencoder experiments."""
