+++ src/ssl4polyp/models/mae/util/crop.py
