+++ tests/test_build_variants.py
@@ -2,17 +2,13 @@ import csv
 import math
 import random
 from pathlib import Path
-import sys
-import pathlib
-
-sys.path.insert(0, str(pathlib.Path(__file__).resolve().parents[1]))
 
 import pytest
 
 cv2 = pytest.importorskip("cv2")
 import numpy as np
 
-from polypdb.robustness.build_variants import (
+from ssl4polyp.polypdb.robustness.build_variants import (
     add_occlusion,
     adjust_brightness,
     adjust_contrast,
